% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolayer.R
\name{autolayer}
\alias{autolayer}
\title{Return ggplot layer objects to overlay prophet plot.}
\usage{
autolayer(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Other arguments passed on to layers.}
}
\value{
A list of layer objects.
}
\description{
Return ggplot layer objects to overlay prophet plot.
}
\examples{
\dontrun{
cpts <- prophet_pick_changepoints(m)
plot(m, fcst) + autolayer(cpts)

outliers <- prophet_detect_outliers(m)
plot(m, fcst) + autolayer(outliers)
}

}
