% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_plot.R
\name{calendar_plot.prophet}
\alias{calendar_plot.prophet}
\title{Calendar plot for residuals of prophet model}
\usage{
\method{calendar_plot}{prophet}(
  object,
  begin = NULL,
  end = NULL,
  drop_empty = TRUE,
  plot_na = FALSE,
  colors = c(low = "red", mid = "white", high = "blue", na = "lightgrey"),
  week_start = 7,
  guide = ggplot2::guide_colorbar("Residual")
)
}
\arguments{
\item{object}{prophet model.}

\item{begin}{date or character value.}

\item{fcst}{data.frame.}
}
\description{
Calendar plot for residuals of prophet model
}
