% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_plot.R
\name{calendar_plot.prophet_outlier}
\alias{calendar_plot.prophet_outlier}
\title{Calendar plot for outliers}
\usage{
\method{calendar_plot}{prophet_outlier}(
  object,
  begin = NULL,
  end = NULL,
  drop_empty = TRUE,
  plot_na = TRUE,
  colors = c(low = "red", mid = "white", high = "blue", na = "lightgrey"),
  week_start = 7,
  guide = ggplot2::guide_colorbar("Residual")
)
}
\arguments{
\item{object}{prophet_outlier object.}

\item{begin}{date or character value.}
}
\description{
Calendar plot for outliers
}
