% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_plot.R
\name{calendar_plot}
\alias{calendar_plot}
\title{Calendar plot}
\usage{
calendar_plot(
  object,
  begin = NULL,
  end = NULL,
  drop_empty = TRUE,
  plot_na,
  colors = c(low = "red", mid = "white", high = "blue", na = "lightgrey"),
  week_start = 7,
  guide = ggplot2::guide_colorbar("Residual")
)
}
\arguments{
\item{object}{Object.}

\item{...}{Other arguments.}
}
\description{
Calendar plot
}
