% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals_calendar.R
\name{plot_residuals_calendar}
\alias{plot_residuals_calendar}
\title{Calendar plot}
\usage{
plot_residuals_calendar(
  object,
  begin = NULL,
  end = NULL,
  drop_empty = TRUE,
  show_na,
  colors = c(low = "red", mid = "white", high = "blue", na = "lightgrey"),
  week_start = 7,
  guide = ggplot2::guide_colorbar("Residual")
)
}
\arguments{
\item{object}{Object.}

\item{...}{Other arguments.}
}
\description{
Calendar plot
}
