% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet_gglayer.R
\name{prophet_gglayer}
\alias{prophet_gglayer}
\title{Return ggplot layer objects to overlay prophet plot.}
\usage{
prophet_gglayer(object, color = "#D55E00", ...)
}
\arguments{
\item{object}{Object.}

\item{color}{Color.}

\item{...}{Other arguments passed on to layers.}
}
\value{
A list of layer objects.
}
\description{
Return ggplot layer objects to overlay prophet plot.
}
\examples{
\dontrun{
cpts <- prophet_pick_changepoints(m)
plot(m, fcst) + prophet_gglayer(cpts)

outliers <- prophet_detect_outliers(m)
plot(m, fcst) + prophet_gglayer(outliers)
}

}
