% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet_pick_changepoints.R
\name{prophet_pick_changepoints}
\alias{prophet_pick_changepoints}
\title{Pick changepoints from prophet object}
\usage{
prophet_pick_changepoints(model, thresh = 0.01)
}
\arguments{
\item{model}{Prophet model object.}

\item{thresh}{Double. Default 0.01.}
}
\value{
A data frame consists of changepoints, growth rates and delta (changes in the growth rates).
}
\description{
Pick changepoints from prophet object
}
\examples{
\dontrun{
m <- prophet(df)
prophet_pick_changepoints(m)
}

}
