% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet_shf.R
\name{prophet_shf}
\alias{prophet_shf}
\title{Simulated Historical Forecast}
\usage{
prophet_shf(model, periods, freq = "d", k = 3, overlap = 0.5)
}
\arguments{
\item{model}{Prophet model object.}

\item{periods}{periods}

\item{freq}{freq}

\item{k}{k}

\item{overlap}{overlap}
}
\description{
Simulated Historical Forecast
}
