% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolayer.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autolayer}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autolayer}}}
}}

