% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet_calendar_plot.R
\name{prophet_calendar_plot}
\alias{prophet_calendar_plot}
\title{Calendar plot}
\usage{
prophet_calendar_plot(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Other arguments.}
}
\description{
Calendar plot
}
