% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet_pick_changepoints.R
\name{prophet_pick_changepoints}
\alias{prophet_pick_changepoints}
\title{Pick Cahngepoints from Prophet Object}
\usage{
prophet_pick_changepoints(model, digit = 2)
}
\arguments{
\item{model}{prophet object}

\item{digit}{effective digit. Default 2.}
}
\description{
Pick Cahngepoints from Prophet Object
}
