% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet_shf.R
\name{prophet_shf}
\alias{prophet_shf}
\title{Simulated Historical Forecast}
\usage{
prophet_shf(model, periods, k = 3, overlap = 0.5, measure_func = c("MAPE",
  "MAE"))
}
\arguments{
\item{model}{Prophet model object.}

\item{periods}{Integer.}

\item{k}{Integer.}

\item{overlap}{Numeric.}

\item{measure_func}{Function or character.}
}
\value{
A prophet shf object.
}
\description{
Simulated Historical Forecast
}
